Remove upper version requirement on Python
aw-qt fails to build on Fedora 41, because it requires Python < 3.13 and Fedora uses Python 3.13.
It works fine with the latest version of Python though.

This patch is intended to help with Fedora packaging by assuming that aw-qt works on the latest Python instead
of requiring the packager to test it manually with every major release of Python.